// src/types/notification.types.ts
export type NotificationType = 'error' | 'success' | 'info';

export enum NotificationTypeEnum {
  error = 'error',
  success = 'success',
  info = 'info',
}

export enum NotificationEnum {
  WelcomeBack = 'We\'re glad to see you again!',
  OrderHasBeenCreated = 'Your order has been successfully created and is now being processed!',
  ProductInCart = 'The product is now in your cart!',
}